DestroysGrades action implemented
